requirements.txt
llmio
python-dotenv
re